[WEAV-186] email 설정
