+++ bootstrap/http/src/main/kotlin/com/studentcenter/weave/bootstrap/WeaveHttpApplication.kt
@@ -3,7 +3,7 @@ package com.studentcenter.weave.bootstrap
 import com.studentcenter.weave.application.common.config.ApplicationConfig
 import com.studentcenter.weave.infrastructure.persistence.common.config.PersistenceConfig
 import com.studentcenter.weave.infrastructure.redis.common.config.RedisConfig
-import com.studentcenter.weave.support.email.ses.config.SesConfig
+import com.studentcenter.weave.support.mail.common.config.MailConfig
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.context.annotation.Import
@@ -14,7 +14,7 @@ import org.springframework.context.annotation.Import
         RedisConfig::class,
         PersistenceConfig::class,
         ApplicationConfig::class,
-        SesConfig::class
+        MailConfig::class,
     ]
 )
 class WeaveHttpApplication
