+++ support/mail/build.gradle.kts
@@ -10,9 +10,8 @@ dependencies {
     implementation(project(":support:common"))
     // 메일 TEMPLATE 사용을 위한 thymeleaf
     implementation("org.springframework.boot:spring-boot-starter-thymeleaf")
+    implementation("org.springframework.boot:spring-boot-starter-mail")
 
-    implementation("aws.sdk.kotlin:ses:${Version.AWS_SDK_KOTLIN_SES}") {
-        exclude("com.squareup.okhttp3:okhttp")
-    }
-    implementation("com.squareup.okhttp3:okhttp:5.0.0-alpha.12")
+    // TODO(SES Sandbox): SES Sandbox 해제시 적용
+    // implementation("aws.sdk.kotlin:ses:${Version.AWS_SDK_KOTLIN_SES}")
 }
