--- support/mail/src/main/kotlin/com/studentcenter/weave/support/email/ses/properties/SesProperties.kt
@@ -1,9 +1,0 @@
-package com.studentcenter.weave.support.email.ses.properties
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(value = "aws.ses")
-data class SesProperties(
-    val accessKey: String,
-    val secretKey: String,
-)
