--- support/mail/src/main/kotlin/com/studentcenter/weave/support/email/ses/service/AmazonSimpleEmailService.kt
@@ -1,19 +1,0 @@
-package com.studentcenter.weave.support.email.ses.service
-
-import aws.sdk.kotlin.services.ses.SesClient
-import aws.sdk.kotlin.services.ses.model.SendEmailRequest
-import kotlinx.coroutines.runBlocking
-import org.springframework.stereotype.Service
-
-@Service
-class AmazonSimpleEmailService(
-    private val sesClient: SesClient,
-) {
-
-    suspend fun sendMailAsync(emailRequest: SendEmailRequest) = sesClient.use {
-        it.sendEmail(emailRequest)
-    }
-
-    fun sendMail(emailRequest: SendEmailRequest) = runBlocking { sendMailAsync(emailRequest) }
-
-}
