+++ support/mail/src/main/kotlin/com/studentcenter/weave/support/mail/common/exception/MailExceptionType.kt
@@ -1,0 +1,7 @@
+package com.studentcenter.weave.support.mail.common.exception
+
+import com.studentcenter.weave.support.common.exception.CustomExceptionType
+
+enum class MailExceptionType(override val code: String) : CustomExceptionType {
+    MAIL_CLIENT_EXCEPTION("EMAIL-001"),
+}
