+++ support/mail/src/main/kotlin/com/studentcenter/weave/support/mail/ses/config/SesConfig.kt
@@ -1,0 +1,25 @@
+package com.studentcenter.weave.support.mail.ses.config
+
+// TODO(SES Sandbox): SES Sandbox 해제시 적용
+//@Configuration
+//@ComponentScan(basePackages = ["com.studentcenter.weave.support.email"])
+//@ConfigurationPropertiesScan(basePackages = ["com.studentcenter.weave.support.email"])
+//class SesConfig(
+//    private val sesProperties: SesProperties,
+//) {
+//
+//    @Bean
+//    fun sesClient() = SesClient {
+//        // region 설정
+//        region = "ap-northeast-2"
+//
+//        // aws 인증
+//        credentialsProvider = StaticCredentialsProvider {
+//            accessKeyId = sesProperties.accessKey
+//            secretAccessKey = sesProperties.secretKey
+//        }
+//
+//
+//    }
+//
+//}
