+++ support/mail/src/main/kotlin/com/studentcenter/weave/support/mail/ses/properties/SesProperties.kt
@@ -1,0 +1,8 @@
+package com.studentcenter.weave.support.mail.ses.properties
+
+// TODO(SES Sandbox): SES Sandbox 해제시 적용
+//@ConfigurationProperties(value = "aws.ses")
+//data class SesProperties(
+//    val accessKey: String,
+//    val secretKey: String,
+//)
