+++ support/mail/src/main/resources/application-mail.yaml
@@ -7,7 +7,23 @@ spring:
     check-template-location: true
     cache: false
 
-aws:
-  ses:
-    access-key: ${AWS_SES_ACCESS_KEY:AKIA6ODU77DHAOXAVIOJ}
-    secret-key: ${AWS_SES_SECRET_KEY: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}
+# TODO(SES Sandbox): SES Sandbox 해제시 적용
+#aws:
+#  ses:
+#    access-key: ${AWS_SES_ACCESS_KEY}
+#    secret-key: ${AWS_SES_SECRET_KEY}
+---
+spring:
+  config:
+    activate:
+      on-profile: dev
+  mail:
+    host: smtp.gmail.com
+    port: 587
+    username: ${DEV_EMAIL_USERNAME}
+    password: ${DEV_EMAIL_PASSWORD}
+    properties:
+      mail:
+        smtp:
+          auth: true
+          starttls.enable: true
